# src/DataService/openapi.yaml
openapi: 3.0.0
info:
  title: Smoke Detection Data Service API
  version: 1.0.0
  description: Node.js microservice for storing and retrieving smoke sensor readings.

servers:
  - url: http://localhost:5001
    description: Local development server

components:
  schemas:
    ReadingInput:
      type: object
      required:
        - utc
        - temperatureC
        - humidityPercent
        - eCo2Ppm
        - tVocPpb
        - pressureHpa
        - pm25
        - fireAlarm
      properties:
        utc:
          type: number
          example: 1650001111
        temperatureC:
          type: number
          example: 22.4
        humidityPercent:
          type: number
          example: 45.1
        eCo2Ppm:
          type: number
          example: 650
        tVocPpb:
          type: number
          example: 120
        pressureHpa:
          type: number
          example: 1013.2
        pm25:
          type: number
          example: 7.8
        fireAlarm:
          type: boolean
          example: false

    CreatedReadingResponse:
      allOf:
        - $ref: '#/components/schemas/ReadingInput'
        - type: object
          properties:
            id:
              type: string
              example: 6807f1ab12cd34ef56ab7890
            createdAt:
              type: string
              format: date-time

    StoredReading:
      allOf:
        - $ref: '#/components/schemas/ReadingInput'
        - type: object
          properties:
            id:
              type: string
              example: 6807f1ab12cd34ef56ab7890
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
              nullable: true

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: Ok
        service:
          type: string
          example: DataService
        mongo:
          type: string
          example: Connected

    DeleteResponse:
      type: object
      properties:
        message:
          type: string
          example: Reading deleted successfully.

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid ID format

paths:
  /health:
    get:
      tags:
        - Health
      summary: Check DataService and MongoDB health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '500':
          description: Service or database is unavailable

  /readings:
    post:
      tags:
        - Readings
      summary: Create a new reading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadingInput'
      responses:
        '201':
          description: Reading created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedReadingResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Readings
      summary: Get readings
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
          description: Maximum number of readings to return
      responses:
        '200':
          description: List of readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoredReading'

  /readings/{id}:
    get:
      tags:
        - Readings
      summary: Get reading by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reading found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredReading'
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reading not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Readings
      summary: Update reading by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadingInput'
      responses:
        '200':
          description: Reading updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredReading'
        '400':
          description: Invalid ID or request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reading not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Readings
      summary: Delete reading by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reading deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reading not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'